Refactor errors into flash messages
Refactored most error messages related to validation into flash messages that are displayed as a list at the top of the page when render_template is called. User can close the messages by clicking them, but the plan is probably to add some sort of timer to the messages.

Removed can_read/can_write from schema for now, not sure about implementing separate read/write permissions.

Updated readme and todo