Add more flash messages and modify their style
User now gets various flash messages near the top of the page for example when creating a new thread or when replying to a thread.

Flash messages are now displayed at a fixed location at the top of the page. They are automatically dismissed after a few seconds, but user can click on them to dismiss them as well.